+++ Terra/Assets/Terra/Combat/Projectiles/ProjectileData.cs
@@ -22,6 +22,6 @@ namespace Terra.Combat.Projectiles
         [Min(2)]public float bulletSpeed;
 
         [Tooltip("Damage dealt by the bullet on hit")]
-        [Min(1)]public float bulletDamage;
+        [Min(1)]public int bulletDamage;
     }
 }
