+++ Terra/Assets/Terra/EffectsSystem/Statuses/Data/PoisonStatusData.cs
@@ -8,6 +8,6 @@ namespace Terra.EffectsSystem.Statuses.Data
     [CreateAssetMenu(fileName = "PoisonStatusData_", menuName = "TheBebocks/Statuses/Data/PoisonStatusData")]
     public class PoisonStatusData : TimedStatusData
     {
-        public float damagePerTick;
+        public int damagePerTick;
     }
 }
